Add files via upload
Summarize the project and what problem it was solving.
In this project, we are tasked with coming up with a system that inputs savings and deposits to show how much you can make with a return interest rate.
What did you do particularly well? In this project, I first had an issue with the syntax of the code. Once I figured this out I think I did tremendously working with a new programming language.
Where could you enhance your code? How would these improvements make your code more efficient, secure, and so on? The way I believe I can enhance my code would be to figure out a way to reduce the amount of code needed.
Which pieces of the code did you find most challenging to write, and how did you overcome this? What tools or resources are you adding to your support network? The hardest part of the code was coming up with the different header files.
What skills from this project will be particularly transferable to other projects or course work? The use of the C++ code language will help in future projects requiring the same language.
How did you make this program maintainable, readable, and adaptable? I made this program readable by using the proper header files, and the correct source.cpp file.